Moved purchase algorithm into utils
Updated ./src/catalog.ts
